+++ src/main/java/com/hbm/blocks/machine/MachineAutocrafter.java
@@ -1,0 +1,73 @@
+package com.hbm.blocks.machine;
+
+import com.hbm.blocks.ModBlocks;
+import com.hbm.lib.InventoryHelper;
+import com.hbm.main.MainRegistry;
+import com.hbm.tileentity.machine.TileEntityMachineAutocrafter;
+import com.hbm.tileentity.machine.TileEntityMachineShredder;
+import net.minecraft.block.BlockContainer;
+import net.minecraft.block.material.Material;
+import net.minecraft.block.properties.PropertyInteger;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.Item;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.EnumBlockRenderType;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.EnumHand;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import net.minecraftforge.fml.common.network.internal.FMLNetworkHandler;
+
+import java.util.Random;
+
+public class MachineAutocrafter extends BlockContainer {
+
+    public static final PropertyInteger FACING = PropertyInteger.create("facing", 2, 5);
+
+    private static boolean keepInventory;
+
+    public MachineAutocrafter(Material materialIn, String s) {
+        super(materialIn);
+        this.setUnlocalizedName(s);
+        this.setRegistryName(s);
+//        this.setDefaultState(this.blockState.getBaseState().withProperty(FACING, 2));\
+        ModBlocks.ALL_BLOCKS.add(this);
+    }
+
+    @Override
+    public Item getItemDropped(IBlockState state, Random rand, int fortune) {
+        return Item.getItemFromBlock(ModBlocks.machine_autocrafter);
+    }
+
+    @Override
+    public TileEntity createNewTileEntity(World worldIn, int meta) {
+        return new TileEntityMachineAutocrafter();
+    }
+
+    @Override
+	public boolean onBlockActivated(World world, BlockPos pos, IBlockState state, EntityPlayer player, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {		if(world.isRemote) {
+			return true;
+		} else if(!player.isSneaking()) {
+			TileEntity entity = world.getTileEntity(pos);
+			if(entity instanceof TileEntityMachineAutocrafter) {
+				FMLNetworkHandler.openGui(player, MainRegistry.instance, 0, world, pos.getX(), pos.getY(), pos.getZ());
+			}
+			return true;
+		} else {
+			return false;
+		}
+	}
+
+    @Override
+    public void breakBlock(World worldIn, BlockPos pos, IBlockState state) {
+        if(!keepInventory)
+            InventoryHelper.dropInventoryItems(worldIn, pos, worldIn.getTileEntity(pos));
+        super.breakBlock(worldIn, pos, state);
+    }
+
+    @Override
+    public EnumBlockRenderType getRenderType(IBlockState state) {
+        return EnumBlockRenderType.MODEL;
+    }
+}