+++ changelog.md
@@ -2,6 +2,7 @@
 ```diff
 + added dummy missile, which can be used to test defences like radar, forcefield, etc. does no damage
 + barrels now save their contents when broken
++ broken ahh autocrafter
 ```
 # fixes
 ```diff